LayoutTests/storage/indexeddb/key-type-infinity.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB key types");

function test()
{
    removeVendorPrefixes();

    name = window.location.pathname;
    description = "My Test Database";
    request = evalAndLog("indexedDB.open(name, description)");
    request.onsuccess = openSuccess;
    request.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    db = evalAndLog("db = event.target.result");

    request = evalAndLog("request = db.setVersion('1')");
    request.onsuccess = testGroup1;
    request.onerror = unexpectedErrorCallback;
}

function testGroup1()
{
    deleteAllObjectStores(db);

    objectStore = evalAndLog("db.createObjectStore('foo');");
    debug("test key as infinity");
    request = evalAndLog("request = objectStore.add([], Infinity);");
    request.onsuccess = testGroup2;
    request.onerror = unexpectedErrorCallback;
}

function testGroup2()
{
    debug("test key as negative infinity");
    request = evalAndLog("request = objectStore.add([], -Infinity);");
    request.onsuccess = testGroup3;
    request.onerror = unexpectedErrorCallback;
}

function testGroup3()
{
    finishJSTest();
}


test();

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
