LayoutTests/storage/indexeddb/key-type-infinity-expected.txt

Test IndexedDB key types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
db.createObjectStore('foo');
test key as infinity
request = objectStore.add([], Infinity);
test key as negative infinity
request = objectStore.add([], -Infinity);
PASS successfullyParsed is true

TEST COMPLETE
